Refactor entity.h Functions
This organizes funcitons believed to be originally found in `entity.c`
into an equivalent `entity.h` and includes that in place of stage
implementations or ASM.
