+++ src/st/entity.h
@@ -1,0 +1,14 @@
+// an implementation of entity.c
+#include "../destroy_entity.h"
+#include "../destroy_entities_from_index.h"
+
+void PreventEntityFromRespawning(Entity* entity) {
+    if (entity->entityRoomIndex) {
+        u32 value = (entity->entityRoomIndex - 1);
+        u16 index = value / 32;
+        u16 bit = value % 32;
+        g_entityDestroyed[index] |= 1 << bit;
+    }
+}
+
+#include "animate_entity.h"
