+++ src/st/no3/41C80.c
@@ -7,20 +7,7 @@ INCLUDE_ASM("st/no3/nonmatchings/41C80", TestCollisions);
 
 #include "../entity_red_door.h"
 
-#include "../../destroy_entity.h"
-
-#include "../../destroy_entities_from_index.h"
-
-void PreventEntityFromRespawning(Entity* entity) {
-    if (entity->entityRoomIndex) {
-        u32 value = (entity->entityRoomIndex - 1);
-        u16 index = value / 32;
-        u16 bit = value % 32;
-        g_entityDestroyed[index] |= 1 << bit;
-    }
-}
-
-#include "../animate_entity.h"
+#include "../entity.h"
 
 u8 func_801C4E4C(u8 frames[], Entity* self, u8 arg2) {
     u16 animFrameStart = self->animFrameIdx * 2;
